// backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PaymentStatus = {
  PENDING: 'PENDING',
  SETTLEMENT: 'SETTLEMENT',
  EXPIRE: 'EXPIRE',
  CANCEL: 'CANCEL',
  DENY: 'DENY',
  REFUND: 'REFUND'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const Role = {
  ADMIN: 'ADMIN',
  FAMILY: 'FAMILY',
  NURSE: 'NURSE'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ServiceType = {
  VISIT: 'VISIT',
  LIVE_IN: 'LIVE_IN',
  LIVE_OUT: 'LIVE_OUT'
} as const

export type ServiceType = (typeof ServiceType)[keyof typeof ServiceType]


export const AppointmentStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  ONGOING: 'ONGOING',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type AppointmentStatus = (typeof AppointmentStatus)[keyof typeof AppointmentStatus]


export const ServiceName = {
  MEDIS: 'MEDIS',
  NON_MEDIS: 'NON_MEDIS'
} as const

export type ServiceName = (typeof ServiceName)[keyof typeof ServiceName]